regulate type coercion

+++ frameworks/napi/input_monitor/src/js_input_monitor_module.cpp
@@ -25,8 +25,9 @@ namespace OHOS {
 namespace MMI {
 namespace {
     constexpr OHOS::HiviewDFX::HiLogLabel LABEL = { LOG_CORE, MMI_LOG_DOMAIN, "JsInputMonitorModule" };
-    constexpr uint32_t MAX_STRING_LEN = 1024;
+    constexpr size_t MAX_STRING_LEN = 1024;
 }
+
 static napi_value JsOn(napi_env env, napi_callback_info info)
 {
     MMI_LOGD("Enter");
